pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "py-wsvnc2"
version = "1.1.0"
description = "WebSocket VNC Client."
authors = [{ name = "Matthew Boyer", email = "matthew.boyer@cynnovative.com"},{ name = "Jon Cwik", email = "jon.cwik@asanet.ai" }]
dependencies = [
  "websockets~=16.0",
  "pillow~=12.1.1",
  "pycryptodomex~=3.23.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }

[project.urls]
homepage = "https://github.com/Cynnovative/py-wsvnc"

[project.optional-dependencies]
dev = ["mypy", "pytest>=9.0.2", "ruff>=0.15.7", "websockify>=0.13.0", "types-Pillow>=10.1.0.1", "coverage>=7.13.5"]

[tool.isort]
profile = "black"

[tool.ruff.isort]
known-first-party = ["wsvnc"]

[tool.mypy]
ignore_missing_imports = false
disallow_untyped_calls = false
disallow_untyped_defs = true
disallow_incomplete_defs = true
files = "src/"
exclude = "tests/"
follow_imports = "silent"

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = ["tests/mock_server_tests"]
log_cli = true
log_cli_level = "INFO"

[tool.ruff.pydocstyle]
convention = "numpy"

[tool.ruff]
extend-select = ["D", "I"]
extend-exclude = [
  "_version.py",
  "notebooks",
]
ignore = [
  "D101",
  "D102",
]

[tool.ruff.per-file-ignores]
"tests/**/test_*.py" = ["D"]
